Update Readme for windows installer/code signing
Updated Readme to include some useful information for Windows installer and code signing stuff. This can be pretty confusing to do if not done before and will save future readers time and confusion digging up how to do these steps. 

I also highly recommend updating the manual > https://build-system.fman.io/manual/#codesigning-windows with the information regarding the signtool.exe requirement as I was not aware this was not installed by default or in the system path and needed to be added. I spent way too much time on it and figured this would save someone else new the same unecessary frustrating experience.

Feel free to reword or format if desired. If you would like screenshots of any of the steps let me know can provide them.